Crate csound_sys

Structs

CSOUND_PARAMS

CS_AUDIODEVICE

CS_MIDIDEVICE

CsoundRandMTState_

ORCTOKEN

RTCLOCK_S

Tree

controlChannelHints_s

controlChannelInfo_s

csRtAudioParams

opcodeListEntry

pvsdat_ext

windat_

Enums

CSOUND

Constants

CSFTYPE_AIFC

CSFTYPE_AIFF

CSFTYPE_ATS

CSFTYPE_AU

CSFTYPE_AVR

CSFTYPE_CAF

CSFTYPE_CVANAL

CSFTYPE_EXTRACT_PARMS

CSFTYPE_FLAC

CSFTYPE_FLOATS_BINARY

CSFTYPE_FLOATS_TEXT

CSFTYPE_FTABLES_BINARY

CSFTYPE_FTABLES_TEXT

CSFTYPE_HETRO

CSFTYPE_HETROT

CSFTYPE_HRTF

CSFTYPE_HTK

CSFTYPE_IMAGE_PNG

CSFTYPE_INTEGER_BINARY

CSFTYPE_INTEGER_TEXT

CSFTYPE_IRCAM

CSFTYPE_LADSPA_PLUGIN

CSFTYPE_LORIS

CSFTYPE_LPC

CSFTYPE_MAT4

CSFTYPE_MAT5

CSFTYPE_MIDI_SYSEX

CSFTYPE_MPC2K

CSFTYPE_NIST

CSFTYPE_OGG

CSFTYPE_OPTIONS

CSFTYPE_ORCHESTRA

CSFTYPE_ORC_INCLUDE

CSFTYPE_OTHER_BINARY

CSFTYPE_OTHER_TEXT

CSFTYPE_PAF

CSFTYPE_POSTSCRIPT

CSFTYPE_PVC

CSFTYPE_PVCEX

CSFTYPE_PVF

CSFTYPE_RAW_AUDIO

CSFTYPE_RF64

CSFTYPE_SCORE

CSFTYPE_SCORE_OUT

CSFTYPE_SCOT

CSFTYPE_SCO_INCLUDE

CSFTYPE_SCRIPT_TEXT

CSFTYPE_SD2

CSFTYPE_SDIF

CSFTYPE_SDS

CSFTYPE_SNAPSHOT

CSFTYPE_SOUNDFONT

CSFTYPE_STD_MIDI

CSFTYPE_SVX

CSFTYPE_UNIFIED_CSD

CSFTYPE_UNKNOWN

CSFTYPE_UNKNOWN_AUDIO

CSFTYPE_UNUSED

CSFTYPE_VOC

CSFTYPE_W64

CSFTYPE_WAVE

CSFTYPE_WAVEX

CSFTYPE_WVE

CSFTYPE_XI

CSFTYPE_XSCANU_MATRIX

CSLANGUAGE_AFRIKAANS

CSLANGUAGE_ALBANIAN

CSLANGUAGE_ARABIC

CSLANGUAGE_ARMENIAN

CSLANGUAGE_ASSAMESE

CSLANGUAGE_AZERI

CSLANGUAGE_BASQUE

CSLANGUAGE_BELARUSIAN

CSLANGUAGE_BENGALI

CSLANGUAGE_BULGARIAN

CSLANGUAGE_CATALAN

CSLANGUAGE_CHINESE

CSLANGUAGE_COLUMBIAN

CSLANGUAGE_CROATIAN

CSLANGUAGE_CZECH

CSLANGUAGE_DANISH

CSLANGUAGE_DEFAULT

CSLANGUAGE_DUTCH

CSLANGUAGE_ENGLISH_UK

CSLANGUAGE_ENGLISH_US

CSLANGUAGE_ESTONIAN

CSLANGUAGE_FAEROESE

CSLANGUAGE_FARSI

CSLANGUAGE_FINNISH

CSLANGUAGE_FRENCH

CSLANGUAGE_GEORGIAN

CSLANGUAGE_GERMAN

CSLANGUAGE_GREEK

CSLANGUAGE_GUJARATI

CSLANGUAGE_HEBREW

CSLANGUAGE_HINDI

CSLANGUAGE_HUNGARIAN

CSLANGUAGE_ICELANDIC

CSLANGUAGE_INDONESIAN

CSLANGUAGE_ITALIAN

CSLANGUAGE_JAPANESE

CSLANGUAGE_KANNADA

CSLANGUAGE_KASHMIRI

CSLANGUAGE_KAZAK

CSLANGUAGE_KONKANI

CSLANGUAGE_KOREAN

CSLANGUAGE_LATVIAN

CSLANGUAGE_LITHUANIAN

CSLANGUAGE_MACEDONIAN

CSLANGUAGE_MALAY

CSLANGUAGE_MALAYALAM

CSLANGUAGE_MANIPURI

CSLANGUAGE_MARATHI

CSLANGUAGE_NEPALI

CSLANGUAGE_NORWEGIAN

CSLANGUAGE_ORIYA

CSLANGUAGE_POLISH

CSLANGUAGE_PORTUGUESE

CSLANGUAGE_PUNJABI

CSLANGUAGE_ROMANIAN

CSLANGUAGE_RUSSIAN

CSLANGUAGE_SANSKRIT

CSLANGUAGE_SERBIAN

CSLANGUAGE_SINDHI

CSLANGUAGE_SLOVAK

CSLANGUAGE_SLOVENIAN

CSLANGUAGE_SPANISH

CSLANGUAGE_SWAHILI

CSLANGUAGE_SWEDISH

CSLANGUAGE_TAMIL

CSLANGUAGE_TATAR

CSLANGUAGE_TELUGU

CSLANGUAGE_THAI

CSLANGUAGE_TURKISH

CSLANGUAGE_UKRAINIAN

CSLANGUAGE_URDU

CSLANGUAGE_UZBEK

CSLANGUAGE_VIETNAMESE

CSOUNDCFG_BOOLEAN

CSOUNDCFG_DOUBLE

CSOUNDCFG_FLOAT

CSOUNDCFG_INTEGER

CSOUNDCFG_INVALID_BOOLEAN

CSOUNDCFG_INVALID_FLAG

CSOUNDCFG_INVALID_NAME

CSOUNDCFG_INVALID_TYPE

CSOUNDCFG_LASTERROR

CSOUNDCFG_MEMORY

CSOUNDCFG_MYFLT

CSOUNDCFG_NOT_POWOFTWO

CSOUNDCFG_NULL_POINTER

CSOUNDCFG_POWOFTWO

CSOUNDCFG_STRING

CSOUNDCFG_STRING_LENGTH

CSOUNDCFG_SUCCESS

CSOUNDCFG_TOO_HIGH

CSOUNDCFG_TOO_LOW

CSOUNDINIT_NO_ATEXIT

CSOUNDINIT_NO_SIGNAL_HANDLER

CSOUNDMSG_BG_BLACK

CSOUNDMSG_BG_BLUE

CSOUNDMSG_BG_COLOR_MASK

CSOUNDMSG_BG_CYAN

CSOUNDMSG_BG_GREEN

CSOUNDMSG_BG_GREY

CSOUNDMSG_BG_MAGENTA

CSOUNDMSG_BG_ORANGE

CSOUNDMSG_BG_RED

CSOUNDMSG_DEFAULT

CSOUNDMSG_ERROR

CSOUNDMSG_FG_ATTR_MASK

CSOUNDMSG_FG_BLACK

CSOUNDMSG_FG_BLUE

CSOUNDMSG_FG_BOLD

CSOUNDMSG_FG_COLOR_MASK

CSOUNDMSG_FG_CYAN

CSOUNDMSG_FG_GREEN

CSOUNDMSG_FG_MAGENTA

CSOUNDMSG_FG_RED

CSOUNDMSG_FG_UNDERLINE

CSOUNDMSG_FG_WHITE

CSOUNDMSG_FG_YELLOW

CSOUNDMSG_ORCH

CSOUNDMSG_REALTIME

CSOUNDMSG_STDOUT

CSOUNDMSG_TYPE_MASK

CSOUNDMSG_WARNING

CSOUND_AUDIO_CHANNEL

CSOUND_CALLBACK_KBD_EVENT

CSOUND_CALLBACK_KBD_TEXT

CSOUND_CHANNEL_TYPE_MASK

CSOUND_CONTROL_CHANNEL

CSOUND_CONTROL_CHANNEL_EXP

CSOUND_CONTROL_CHANNEL_INT

CSOUND_CONTROL_CHANNEL_LIN

CSOUND_CONTROL_CHANNEL_NO_HINTS

CSOUND_ERROR

CSOUND_EXITJMP_SUCCESS

CSOUND_INITIALIZATION

CSOUND_INPUT_CHANNEL

CSOUND_MEMORY

CSOUND_OUTPUT_CHANNEL

CSOUND_PERFORMANCE

CSOUND_PVS_CHANNEL

CSOUND_SIGNAL

CSOUND_STRING_CHANNEL

CSOUND_SUCCESS

CSOUND_VAR_CHANNEL

Functions

csoundAddSpinSample⚠

csoundAppendOpcode⚠

csoundCfgErrorCodeToString⚠

csoundCleanup⚠

csoundClearSpin⚠

csoundCloseLibrary⚠

csoundCompile⚠

csoundCompileArgs⚠

csoundCompileCsd⚠

csoundCompileCsdText⚠

csoundCompileOrc⚠

csoundCompileOrcAsync⚠

csoundCompileTree⚠

csoundCreate⚠

csoundCreateBarrier⚠

csoundCreateCircularBuffer⚠

csoundCreateConfigurationVariable⚠

csoundCreateGlobalVariable⚠

csoundCreateMessageBuffer⚠

csoundCreateMutex⚠

csoundCreateThread⚠

csoundCreateThreadLock⚠

csoundDeleteChannelList⚠

csoundDeleteTree⚠

csoundDeleteUtilityList⚠

csoundDestroy⚠

csoundDestroyBarrier⚠

csoundDestroyCircularBuffer⚠

csoundDestroyGlobalVariable⚠

csoundDestroyMessageBuffer⚠

csoundDestroyMutex⚠

csoundDestroyThreadLock⚠

csoundDisposeOpcodeList⚠

csoundEvalCode⚠

csoundFlushCircularBuffer⚠

csoundGet0dBFS⚠

csoundGetA4⚠

csoundGetAPIVersion⚠

csoundGetAudioChannel⚠

csoundGetAudioDevList⚠

csoundGetCPUTime⚠

csoundGetChannelDatasize⚠

csoundGetChannelLock⚠

csoundGetChannelPtr⚠

csoundGetControlChannel⚠

csoundGetControlChannelHints⚠

csoundGetCurrentThreadId⚠

csoundGetCurrentTimeSamples⚠

csoundGetDebug⚠

csoundGetEnv⚠

csoundGetFirstMessage⚠

csoundGetFirstMessageAttr⚠

csoundGetHostData⚠

csoundGetInputBuffer⚠

csoundGetInputBufferSize⚠

csoundGetInputName⚠

csoundGetKr⚠

csoundGetKsmps⚠

csoundGetLibrarySymbol⚠

csoundGetMIDIDevList⚠

csoundGetMessageCnt⚠

csoundGetMessageLevel⚠

csoundGetModule⚠

csoundGetNamedGEN⚠

csoundGetNamedGens⚠

csoundGetNchnls⚠

csoundGetNchnlsInput⚠

csoundGetOutputBuffer⚠

csoundGetOutputBufferSize⚠

csoundGetOutputFormat⚠

csoundGetOutputName⚠

csoundGetParams⚠

csoundGetPvsChannel⚠

csoundGetRandomSeedFromTime⚠

csoundGetRealTime⚠

csoundGetRtPlayUserData⚠

csoundGetRtRecordUserData⚠

csoundGetScoreOffsetSeconds⚠

csoundGetScoreTime⚠

csoundGetSizeOfMYFLT⚠

csoundGetSpin⚠

csoundGetSpout⚠

csoundGetSpoutSample⚠

csoundGetSr⚠

csoundGetStringChannel⚠

csoundGetTable⚠

csoundGetTableArgs⚠

csoundGetUtilityDescription⚠

csoundGetVersion⚠

csoundInitTimerStruct⚠

csoundInitialize⚠

csoundInputMessage⚠

csoundInputMessageAsync⚠

csoundIsNamedGEN⚠

csoundIsScorePending⚠

csoundJoinThread⚠

csoundKeyPress⚠

csoundKillInstance⚠

csoundListChannels⚠

csoundListUtilities⚠

csoundLocalizeString⚠

csoundLockMutex⚠

csoundLockMutexNoWait⚠

csoundMessage⚠

csoundNewOpcodeList⚠

csoundNotifyThreadLock⚠

csoundOpenLibrary⚠

csoundParseConfigurationVariable⚠

csoundParseOrc⚠

csoundPeekCircularBuffer⚠

csoundPerform⚠

csoundPerformBuffer⚠

csoundPerformKsmps⚠

csoundPopFirstMessage⚠

csoundQueryGlobalVariable⚠

csoundQueryGlobalVariableNoCheck⚠

csoundRand31⚠

csoundRandMT⚠

csoundReadCircularBuffer⚠

csoundReadScore⚠

csoundReadScoreAsync⚠

csoundRegisterKeyboardCallback⚠

csoundRegisterSenseEventCallback⚠

csoundRemoveKeyboardCallback⚠

csoundReset⚠

csoundRewindScore⚠

csoundRunCommand⚠

csoundRunUtility⚠

csoundScoreEvent⚠

csoundScoreEventAbsolute⚠

csoundScoreEventAbsoluteAsync⚠

csoundScoreEventAsync⚠

csoundScoreExtract⚠

csoundScoreSort⚠

csoundSeedRandMT⚠

csoundSetAudioChannel⚠

csoundSetAudioDeviceListCallback⚠

csoundSetConfigurationVariable⚠

csoundSetControlChannel⚠

csoundSetControlChannelHints⚠

csoundSetCscoreCallback⚠

csoundSetDebug⚠

csoundSetDrawGraphCallback⚠

csoundSetExitGraphCallback⚠

csoundSetExternalMidiErrorStringCallback⚠

csoundSetExternalMidiInCloseCallback⚠

csoundSetExternalMidiInOpenCallback⚠

csoundSetExternalMidiOutCloseCallback⚠

csoundSetExternalMidiOutOpenCallback⚠

csoundSetExternalMidiReadCallback⚠

csoundSetExternalMidiWriteCallback⚠

csoundSetFileOpenCallback⚠

csoundSetGlobalEnv⚠

csoundSetHostData⚠

csoundSetHostImplementedAudioIO⚠

csoundSetHostImplementedMIDIIO⚠

csoundSetInput⚠

csoundSetInputChannelCallback⚠

csoundSetIsGraphable⚠

csoundSetKillGraphCallback⚠

csoundSetLanguage⚠

csoundSetMIDIDeviceListCallback⚠

csoundSetMIDIFileInput⚠

csoundSetMIDIFileOutput⚠

csoundSetMIDIInput⚠

csoundSetMIDIModule⚠

csoundSetMIDIOutput⚠

csoundSetMakeGraphCallback⚠

csoundSetMessageLevel⚠

csoundSetMessageStringCallback⚠

csoundSetOption⚠

csoundSetOutput⚠

csoundSetOutputChannelCallback⚠

csoundSetParams⚠

csoundSetPlayopenCallback⚠

csoundSetPvsChannel⚠

csoundSetRTAudioModule⚠

csoundSetRecopenCallback⚠

csoundSetRtcloseCallback⚠

csoundSetRtplayCallback⚠

csoundSetRtrecordCallback⚠

csoundSetScoreOffsetSeconds⚠

csoundSetScorePending⚠

csoundSetSpinSample⚠

csoundSetStringChannel⚠

csoundSetYieldCallback⚠

csoundSleep⚠

csoundStart⚠

csoundStop⚠

csoundStopUDPConsole⚠

csoundTableCopyIn⚠

csoundTableCopyInAsync⚠

csoundTableCopyOut⚠

csoundTableCopyOutAsync⚠

csoundTableGet⚠

csoundTableLength⚠

csoundTableSet⚠

csoundUDPConsole⚠

csoundUDPServerClose⚠

csoundUDPServerStart⚠

csoundUDPServerStatus⚠

csoundUnlockMutex⚠

csoundWaitBarrier⚠

csoundWaitThreadLock⚠

csoundWaitThreadLockNoTimeout⚠

csoundWriteCircularBuffer⚠

Type Definitions

CSOUND_FILETYPES_t
The following constants are used with csound->FileOpen2() and csound->ldmemfile2() to specify the format of a file that is being opened. This information is passed by Csound to a host's FileOpen callback and does not influence the opening operation in any other way. Conversion from Csound's TYP_XXX macros for audio formats to CSOUND_FILETYPES values can be done with csound->type2csfiletype().

CSOUND_STATUS

CsoundRandMTState

PVSDATEXT

RTCLOCK

WINDAT

controlChannelBehavior

controlChannelHints_t

controlChannelInfo_t

controlChannelType

csLenguage_t

cscore_callback_type

csound_channel_callback

csound_dev_list_callback

csound_ext_midi_close_callback

csound_ext_midi_error_callback

csound_ext_midi_open_callback

csound_ext_midi_read_data_callback

csound_ext_midi_write_data_callback

csound_file_open_callback

csound_message_callback

csound_midi_dev_list_callback

csound_open_callback

csound_rt_close_callback

csound_rt_play_callback

csound_rt_rec_callback